# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Echobot",
    "description": "AI驱动的数字人动画平台 API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/voice-clone/preview": {
      "post": {
        "tags": [
          "语音克隆"
        ],
        "summary": "创建语音克隆预览任务",
        "description": "创建语音克隆预览任务\n\n上传家长的参考音频，系统会克隆音色并用该音色朗读选定的预设故事。\n这是一个异步任务，创建后需要轮询 GET /preview/{task_id} 获取结果。\n\n**音频要求:**\n- 时长：10-15秒\n- 格式：WAV 或 MP3\n- 内容：清晰的语音，无背景噪音\n- 大小：最大10MB\n\n**请求参数:**\n- **audio**: 参考音频文件（multipart/form-data）\n- **story_id**: 预设故事ID\n\n**返回示例:**\n```json\n{\n    \"code\": 0,\n    \"message\": \"success\",\n    \"data\": {\n        \"task_id\": \"user123_1703980800000\",\n        \"status\": \"processing\",\n        \"message\": \"语音克隆任务已创建，请轮询获取状态\"\n    }\n}\n```",
        "operationId": "create_voice_clone_preview_api_v1_voice_clone_preview_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_voice_clone_preview_api_v1_voice_clone_preview_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/voice-clone/preview/{task_id}": {
      "get": {
        "tags": [
          "语音克隆"
        ],
        "summary": "获取语音克隆任务状态",
        "description": "获取语音克隆任务状态\n\n轮询此接口获取任务的处理进度和结果。\n建议轮询间隔：2-3秒。\n\n**路径参数:**\n- **task_id**: 任务ID（从 POST /preview 返回）\n\n**任务状态说明:**\n- **processing**: 处理中，继续轮询\n- **completed**: 完成，audio_url 包含生成的音频\n- **failed**: 失败，error 包含错误信息\n\n**返回示例:**\n```json\n{\n    \"code\": 0,\n    \"message\": \"success\",\n    \"data\": {\n        \"task_id\": \"user123_1703980800000\",\n        \"status\": \"completed\",\n        \"progress\": 100,\n        \"audio_url\": \"/uploads/voice/user123/preview.mp3\",\n        \"error\": null\n    }\n}\n```",
        "operationId": "get_voice_clone_status_api_v1_voice_clone_preview__task_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/voice-clone/profiles": {
      "get": {
        "tags": [
          "语音克隆"
        ],
        "summary": "获取声音档案列表",
        "description": "获取用户的所有声音档案\n\n返回当前用户保存的所有声音档案列表。\n这些声音可用于故事生成、有声书等功能。\n\n**返回示例:**\n```json\n{\n    \"code\": 0,\n    \"message\": \"success\",\n    \"data\": {\n        \"profiles\": [\n            {\n                \"id\": \"694534f1c32ffbd6151c18a3\",\n                \"name\": \"爸爸的声音\",\n                \"voice_id\": \"cosyvoice-clone-xxxx\",\n                \"reference_audio_url\": \"/uploads/voice/user123/ref.wav\",\n                \"preview_audio_url\": \"/uploads/voice/user123/preview.mp3\",\n                \"created_at\": \"2025-01-01T12:00:00\"\n            }\n        ],\n        \"total\": 2\n    }\n}\n```",
        "operationId": "get_voice_profiles_api_v1_voice_clone_profiles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "语音克隆"
        ],
        "summary": "保存声音档案",
        "description": "保存声音档案\n\n语音克隆预览完成且满意后，将克隆的声音保存到声音档案库。\n保存后可在故事生成、有声书等功能中重复使用该声音。\n\n**请求参数:**\n- **task_id**: 已完成的语音克隆任务ID\n- **name**: 声音名称，如：爸爸的声音、妈妈的声音\n\n**返回示例:**\n```json\n{\n    \"code\": 0,\n    \"message\": \"success\",\n    \"data\": {\n        \"profile\": {\n            \"id\": \"694534f1c32ffbd6151c18a3\",\n            \"name\": \"爸爸的声音\",\n            \"voice_id\": \"cosyvoice-clone-xxxx\",\n            \"reference_audio_url\": \"/uploads/voice/user123/ref.wav\",\n            \"preview_audio_url\": \"/uploads/voice/user123/preview.mp3\",\n            \"created_at\": \"2025-01-01T12:00:00\"\n        }\n    }\n}\n```",
        "operationId": "save_voice_profile_api_v1_voice_clone_profiles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveVoiceProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/voice-clone/profiles/{profile_id}": {
      "get": {
        "tags": [
          "语音克隆"
        ],
        "summary": "获取声音档案详情",
        "description": "获取单个声音档案详情\n\n**路径参数:**\n- **profile_id**: 声音档案ID\n\n**返回示例:**\n```json\n{\n    \"code\": 0,\n    \"message\": \"success\",\n    \"data\": {\n        \"id\": \"694534f1c32ffbd6151c18a3\",\n        \"name\": \"爸爸的声音\",\n        \"voice_id\": \"cosyvoice-clone-xxxx\",\n        \"reference_audio_url\": \"/uploads/voice/user123/ref.wav\",\n        \"preview_audio_url\": \"/uploads/voice/user123/preview.mp3\",\n        \"created_at\": \"2025-01-01T12:00:00\"\n    }\n}\n```",
        "operationId": "get_voice_profile_api_v1_voice_clone_profiles__profile_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "语音克隆"
        ],
        "summary": "更新声音档案",
        "description": "更新声音档案名称\n\n**路径参数:**\n- **profile_id**: 声音档案ID\n\n**请求参数:**\n- **name**: 新的声音名称\n\n**返回示例:**\n```json\n{\n    \"code\": 0,\n    \"message\": \"更新成功\",\n    \"data\": null\n}\n```",
        "operationId": "update_voice_profile_api_v1_voice_clone_profiles__profile_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVoiceProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "语音克隆"
        ],
        "summary": "删除声音档案",
        "description": "删除声音档案\n\n删除后将无法恢复，且使用该声音的历史记录不受影响。\n\n**路径参数:**\n- **profile_id**: 声音档案ID\n\n**返回示例:**\n```json\n{\n    \"code\": 0,\n    \"message\": \"删除成功\",\n    \"data\": null\n}\n```",
        "operationId": "delete_voice_profile_api_v1_voice_clone_profiles__profile_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_voice_clone_preview_api_v1_voice_clone_preview_post": {
        "properties": {
          "audio": {
            "type": "string",
            "format": "binary",
            "title": "Audio",
            "description": "参考音频文件，10-15秒清晰语音，WAV/MP3格式"
          },
          "story_id": {
            "type": "string",
            "title": "Story Id",
            "description": "预设故事ID，从 /preset-stories 获取"
          }
        },
        "type": "object",
        "required": [
          "audio",
          "story_id"
        ],
        "title": "Body_create_voice_clone_preview_api_v1_voice_clone_preview_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SaveVoiceProfileRequest": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id",
            "description": "已完成的语音克隆任务ID",
            "example": "user123_1703980800000"
          },
          "name": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Name",
            "description": "声音名称，便于识别，如：爸爸的声音、妈妈的声音",
            "example": "爸爸的声音"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "name"
        ],
        "title": "SaveVoiceProfileRequest",
        "description": "保存声音档案请求\n\n语音克隆预览成功后，将克隆的声音保存为档案。\n保存后可在故事生成、有声书等功能中重复使用。",
        "example": {
          "name": "爸爸的声音",
          "task_id": "user123_1703980800000"
        }
      },
...
      "UpdateVoiceProfileRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Name",
            "description": "新的声音名称",
            "example": "妈妈的声音"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "UpdateVoiceProfileRequest",
        "description": "更新声音档案请求\n\n用于修改声音档案的名称。",
        "example": {
          "name": "妈妈的声音"
        }
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}